This is simple CRUD web application used for testing potential employees.
## Story #1

Add ability to manage clients (companies):

* The user should be able to create, edit, delete and list Clients.
* Clients should have a company name, website URI, phone number, and
physical/mailing address.
* Clients can have zero, one, or multiple associated contacts.
* When editing a person, the user should be able to choose the
associated client.
* When viewing a person, the associated client should be shown.
* When viewing a client, the associated contacts should be shown.
* When editing a client, the user should be able to add or remove
associated contacts.

## Story #2

Add client­side styling and validation:

* Using jQuery and your preferred javascript validation technique,
prevent the submission on invalid data on the edit forms and inform the
user of errors. The validations implemented on the front­end should be
equivalent to the existing server side validation.
* Add standard styling to the pages using a modern CSS3/HTML5 framework
like Twitter Bootstrap or Foundation (or similar).

The specification implied that there is a zero, one-to-many
relationship between clients and contacts and a zero, one to one
relationship between contacts and clients.
These relationships are reflected in the app. A contact will only be
associated with at most one client. It is possible to have clients that
are not associated with any
contacts and vice versa.